DEV-93 - good JSON serialization (#18)
Made good JSON serialization for model objects.
