+++ etg/__version__.py
@@ -1,7 +1,7 @@
 __title__ = 'etg'
 __description__ = 'ETG API v3 Python SDK'
 __url__ = ''
-__version__ = '0.9.1'
+__version__ = '0.10.0'
 __author__ = 'Sergey Popinevskiy'
 __author_email__ = 'sergey.popinevskiy@gmail.com'
 __license__ = 'MIT'
